Create README.md
# TaskTrackingSolution
This Task Tracking website will be used by a fictitious Organization and will allow employees of an organization to create a new project and add user stories. Once the user stories are in place, the project manager will assign the tasks to  team members. Team members will perform the task and update the progress of the task(s) on the web site. After this, the project manager can review the tasks online and provide comments on incomplete tasks